refactor(ludo): Update LudoPlayer to extend Player abstract for unified game model
